+++ src/router/index.ts
@@ -3,7 +3,7 @@ import HomeView from '@/pages/HomeView.vue'
 import NotFoundView from '@/pages/NotFoundView.vue'
 
 const router = createRouter({
-  history: createWebHistory(),
+  history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     { path: '/', name: 'home', component: HomeView },
     // ADD ALL CUSTOM ROUTES ABOVE THE CATCH-ALL ROUTE